Shopping Cart Program
Let’s write a Python program to implement a shopping cart price calculator. Our store sells four items and each item costs $1. Currently our store is offering some promotions. These are the promotions applied to each of the four items:

Item A = Full price only
Item B = Buy one get one free
Item C = 50% off
Item D = Full price only

Write a function or program that will calculate the total for a shopping cart based on the given prices and the given promotions. For example:

If the shopping cart contained “A, B, C, D” then the total would be $1 + $1 + $0.50 + $1 = $3.50 because Item C is 50% and even though Item B is buy one get one free there was only one item in the cart.

If the shopping cart contained “A, B, B, D” then the total would be $1 + $1 + $1 - $1 + $1 = $3 because Item B appeared in the cart twice so the second one was free.